% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat.plagiarism.R
\name{as.list.textstat_plagiarism}
\alias{as.list.textstat_plagiarism}
\title{Returns reused text segments}
\usage{
\method{as.list}{textstat_plagiarism}(x, upper = TRUE, ...)
}
\arguments{
\item{x}{\code{textstat_plagiarism} object}

\item{upper}{if \code{TRUE}, cosider upper-triangular pairs. Set \code{FALSE}
when documents are in chronological order.}

\item{...}{not used}
}
\description{
Returns reused text segments
}
