% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat.plagiarism.R
\name{textstat_plagiarism}
\alias{textstat_plagiarism}
\title{A function to detects reused text segments}
\usage{
textstat_plagiarism(x, y = NULL, size = 10, min_simil = 0.9)
}
\arguments{
\item{x}{a \link[quanteda]{tokens} object in which texts are resued}

\item{y}{a \link[quanteda]{tokens} object from which texts are taken}

\item{size}{size of the reused text segments specified in number of tokens}

\item{min_simil}{minimum similarity of text segments that are treated as reuse}
}
\description{
This is an example of how to make a custom function based on quanteda.
}
\examples{

require(quanteda)
quanteda_options(threads = 8)
corp <- data_corpus_inaugural
toks <- tokens(corp, remove_punct = TRUE) \%>\%
    tokens_remove(stopwords("en")) \%>\%
    tokens_select("^[\\\\p{L}]+$", valuetype = "regex")

# unigram
tstat_plag <- textstat_plagiarism(
    toks[1:10],
    min_simil = 0.5
)
print(tstat_plag)
as.list(tstat_plag, upper = FALSE) \%>\% head()
as.matrix(tstat_plag, upper = FALSE)

# bigram
tstat_plag2 <- textstat_plagiarism(
    tokens_ngrams(toks, 2),
    size = 5,
    min_simil = 0.5
)

print(tstat_plag2)
as.list(tstat_plag2, upper = FALSE) \%>\% head()
as.matrix(tstat_plag2)
}
